requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.6
customtkinter==5.2.2
docutils==0.18.1
docutils==0.18.1
filelock==3.18.0
HTMLParser==0.0.2
ipython==8.30.0
ipython==9.1.0
ipython==8.12.3
ipywidgets==7.8.1
jnius==1.1.0
joblib==1.4.2
keyring==24.3.1
matplotlib==3.10.5
numpy==2.3.2
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
opencv_python_headless==4.12.0.88
pandas==2.3.1
Pillow==11.3.0
protobuf==4.25.3
pyOpenSSL==24.2.1
pyOpenSSL==25.1.0
redis==6.3.0
scikit_learn==1.7.1
seaborn==0.13.2
Sphinx==8.2.3
tensorflow==2.18.0
tensorflow_hub==0.16.1
tensorflow_intel==2.18.0
thread==2.0.5
umap==0.1.1
umap_learn==0.5.7
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
